requirements.txt
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
passlib
python-decouple
python-jose
python-multipart
sqlalchemy
alembic
psycopg2
celery==4.4.7
pika==1.2.0